requirements.txt
pandas
numpy
opencv-python-headless
pymongo
python-dotenv